refactor(ui): split workspace conflict rendering
